set up index and server files update gitignore remove redundant files
