+++ .gitignore
@@ -78,3 +78,6 @@ typings/
 
 # DynamoDB Local files
 .dynamodb/
+.env
+*.db3
+.vscode/
