--- api/server.js
@@ -1,15 +1,0 @@
-const express = require('express');
-const cors = require('cors');
-const helmet = require('helmet');
-
-const configureRoutes = require('../config/routes.js');
-
-const server = express();
-
-server.use(helmet());
-server.use(cors());
-server.use(express.json());
-
-configureRoutes(server);
-
-module.exports = server;
