+++ index.js
@@ -1,8 +1,9 @@
 require('dotenv').config(); // load .env variables
 
-const server = require('./api/server.js');
+const server = require('./server');
 
-const port = process.env.PORT || 3300;
+const port = 3300;
+
 server.listen(port, () => {
   console.log(`\n=== Server listening on port ${port}\n`);
 });
